+++ packages/react-testing/CHANGELOG.md
@@ -9,7 +9,7 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ### Fixed
 
-- `useMountedRef` now works with React 17 [#1964](https://github.com/Shopify/quilt/pull/1964).
+- Changes to tests for compatibility in dev. [#1957](https://github.com/Shopify/quilt/issues/1957)
 
 ## 3.1.0 - 2021-06-29